uwuzu v1.6.7 Hapuego

+++ settings_admin/api/update_query.php
@@ -25,61 +25,68 @@ if (isset($_FILES['update_zip']) && isset($_POST['userid']) && isset($_POST['acc
         );
         $pdo = new PDO('mysql:charset=utf8mb4;dbname=' . DB_NAME . ';host=' . DB_HOST, DB_USER, DB_PASS, $option);
     } catch (PDOException $e) {
-        // 接続エラーのときエラー内容を取得する
         $error_message[] = $e->getMessage();
     }
 
-    $query = $pdo->prepare('SELECT * FROM account WHERE userid = :userid limit 1');
+    if(empty($error_message)){
+        $query = $pdo->prepare('SELECT * FROM account WHERE userid = :userid limit 1');
+        $query->execute(array(':userid' => $postUserid));
+        $result2 = $query->fetch();
 
-    $query->execute(array(':userid' => $postUserid));
+        if($result2["loginid"] === $loginid){
+            if($result2["admin"] === "yes"){
 
-    $result2 = $query->fetch();
+                $uploadDir = sys_get_temp_dir();
+                $uploadFile = $uploadDir . '/' . basename($_FILES['update_zip']['name']);
 
-    if($result2["loginid"] === $loginid){
-        if($result2["admin"] === "yes"){
+                if (move_uploaded_file($_FILES['update_zip']['tmp_name'], $uploadFile)) {
+                    $extractPath = $uploadDir . '/uwuzu_update_' . createUniqId();
 
-            $uploadDir = sys_get_temp_dir();
-            $uploadFile = $uploadDir . '/' . basename($_FILES['update_zip']['name']);
+                    $zip = new ZipArchive;
+                    if ($zip->open($uploadFile) == true) {
+                        $zip->extractTo($extractPath);
+                        $zip->close();
 
-            if (move_uploaded_file($_FILES['update_zip']['tmp_name'], $uploadFile)) {
-                $extractPath = $uploadDir . '/uwuzu_update_' . createUniqId();
-
-                $zip = new ZipArchive;
-                if ($zip->open($uploadFile) == true) {
-                    $zip->extractTo($extractPath);
-                    $zip->close();
-
-                    // JSONファイルを読み込む
-                    $jsonFile = $extractPath . '/update.json';
-                    if (file_exists($jsonFile)) {
-                        $jsonData = json_decode(file_get_contents($jsonFile), true);
-                        if (json_last_error() === JSON_ERROR_NONE) {
-                            $response = [
-                                'success' => true,
-                                'software_name' => safetext($jsonData['software']) ?? '名前がありません',
-                                'version' => safetext($jsonData['version']) ?? 'バージョン情報がありません',
-                                'release_notes' => safetext($jsonData['release_notes']) ?? 'リリースノートが見つかりません。',
-                                'notices' => safetext($jsonData['notices']) ?? '注意事項が見つかりません。',
-                                'file_path' => safetext($extractPath)
-                            ];
-                            echo json_encode($response);
+                        // JSONファイルを読み込む
+                        $jsonFile = $extractPath . '/update.json';
+                        if (file_exists($jsonFile)) {
+                            $jsonData = json_decode(file_get_contents($jsonFile), true);
+                            if (json_last_error() === JSON_ERROR_NONE) {
+                                $response = [
+                                    'success' => true,
+                                    'software_name' => safetext($jsonData['software']) ?? '名前がありません',
+                                    'version' => safetext($jsonData['version']) ?? 'バージョン情報がありません',
+                                    'release_notes' => safetext($jsonData['release_notes']) ?? 'リリースノートが見つかりません。',
+                                    'notices' => safetext($jsonData['notices']) ?? '注意事項が見つかりません。',
+                                    'file_path' => safetext($extractPath)
+                                ];
+                                echo json_encode($response);
+                            } else {
+                                echo json_encode(['success' => false, 'error' => 'JSONファイルの読み込みに失敗しました。(ROADING_JSON_ERROR)']);
+                            }
                         } else {
                             echo json_encode(['success' => false, 'error' => 'JSONファイルの読み込みに失敗しました。(ROADING_JSON_ERROR)']);
                         }
                     } else {
-                        echo json_encode(['success' => false, 'error' => 'JSONファイルの読み込みに失敗しました。(ROADING_JSON_ERROR)']);
+                        echo json_encode(['success' => false, 'error' => '読み込みに失敗しました。1(ROADING_ERROR)']);
                     }
-                } else {
-                    echo json_encode(['success' => false, 'error' => '読み込みに失敗しました。1(ROADING_ERROR)']);
-                }
 
-                if (file_exists($uploadFile)) {
-                    if (is_file($uploadFile)) {
-                        unlink($uploadFile);
+                    if (file_exists($uploadFile)) {
+                        if (is_file($uploadFile)) {
+                            unlink($uploadFile);
+                        }
                     }
+                }else{
+                    echo json_encode(['success' => false, 'error' => 'ファイルの移動に失敗しました。(ROADING_ERROR)']);
                 }
+            }else{
+                echo json_encode(['success' => false, 'error' => '認証に失敗しました。(AUTH_INVALID)']);
             }
+        }else{
+            echo json_encode(['success' => false, 'error' => '認証に失敗しました。(AUTH_INVALID)']);
         }
+    }else{
+        echo json_encode(['success' => false, 'error' => 'データベースの接続に失敗しました。(ERROR)']);
     }
 }else{
     echo json_encode(['success' => false, 'error' => '読み込みに失敗しました。2(ROADING_ERROR)']);
